Refactored `HasGraduated`
- Return Type bool
Changed to a more appropriate boolean return type instead of an obscure tuple.
Standing is now an out parameter

- Diploma credit check
Seemed to have been a bug. Diploma credit criteria was not being verified.

- Missing SumaCumLaude
Standing was missing SumaCumLaude determiniation and had MangnaCumLaude twice instead

- Increased accuracy
Average is now a decimal so that accuracy is preserved